Fix #55, move parsing out of Todo to Parser.
